Move StatusFromStatusCode to commons dir
