+++ integrations/access/common/status.go
@@ -18,6 +18,7 @@ package common
 
 import (
 	"context"
+	"net/http"
 
 	"github.com/gravitational/teleport/api/types"
 )
@@ -25,4 +26,17 @@ import (
 // StatusSink defines a destination for PluginStatus
 type StatusSink interface {
 	Emit(ctx context.Context, s types.PluginStatus) error
+}
+
+func StatusFromStatusCode(httpCode int) types.PluginStatus {
+	var code types.PluginStatusCode
+	switch {
+	case httpCode == http.StatusUnauthorized:
+		code = types.PluginStatusCode_UNAUTHORIZED
+	case httpCode >= 200 && httpCode < 400:
+		code = types.PluginStatusCode_RUNNING
+	default:
+		code = types.PluginStatusCode_OTHER_ERROR
+	}
+	return &types.PluginStatusV1{Code: code}
 }
